chore(avm)!: make indirects big enough for relative addressing (#9000)

+++ yarn-project/simulator/src/avm/opcodes/external_calls.ts
@@ -15,7 +15,7 @@ abstract class ExternalCall extends Instruction {
   // Informs (de)serialization. See Instruction.deserialize.
   static readonly wireFormat: OperandType[] = [
     OperandType.UINT8,
-    OperandType.UINT8,
+    OperandType.UINT16, // Indirect
     OperandType.UINT32,
     OperandType.UINT32,
     OperandType.UINT32,
